Make Slackish class, refactor Slack emitter to use it
Also fix refactor breakage with quantity fields
